# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RemessasSteam API",
        "description": "API para gestão de projetos, parceiros e divisão de receitas",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/payouts": {
            "get": {
                "tags": [
                    "Payouts"
                ],
                "summary": "Listar payouts",
                "operationId": "0943d4399d51d89d6b212c8ffe31d60e",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "partner_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "project_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "report_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de payouts"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payouts/{payout}": {
            "get": {
                "tags": [
                    "Payouts"
                ],
                "summary": "Detalhar payout",
                "operationId": "9ac4d74b28e723e96d0f2fedf7aab5bc",
                "parameters": [
                    {
                        "name": "payout",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payout"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Payouts"
                ],
                "summary": "Atualizar payout",
                "operationId": "1fd919725a012f3bfb447741120ae472",
                "parameters": [
                    {
                        "name": "payout",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "scheduled",
                                            "paid",
                                            "canceled"
                                        ]
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "method": {
                                        "type": "string",
                                        "enum": [
                                            "pix",
                                            "transfer",
                                            "paypal",
                                            "wise",
                                            "other"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payout atualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Payouts"
                ],
                "summary": "Remover payout",
                "operationId": "4f3ab17c7c27d93d59869f8a125fbbbe",
                "parameters": [
                    {
                        "name": "payout",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Payout removido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/{report}/generate-payouts": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Gerar payouts a partir de relatório",
                "operationId": "f01072445d326c35cf91849cff88935d",
                "parameters": [
                    {
                        "name": "report",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "reset_existing": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payouts gerados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payouts/{payout}/mark-paid": {
            "post": {
                "tags": [
                    "Payouts"
                ],
                "summary": "Marcar payout como pago",
                "operationId": "ca5265fbf15d880d975bec8398e7d350",
                "parameters": [
                    {
                        "name": "payout",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "paid_at": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "method": {
                                        "type": "string",
                                        "enum": [
                                            "pix",
                                            "transfer",
                                            "paypal",
                                            "wise",
                                            "other"
                                        ]
                                    },
                                    "receipt": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "partner_invoice_number": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "partner_invoice": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payout marcado como pago"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}